requirements.txt
selenium
termcolor
datetime
urllib3
requests
python-dotenv